+++ activity/brightpay/python/changelog.md
@@ -10,3 +10,14 @@
 
 ---
 
+# Brightpay Development - 2025-01-04
+
+**Commit #9**
+
+- Repository: brightpay/checkout-microservice
+- Type: microservice
+- Language: python
+- Month: 1/2025
+
+---
+